apps/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic==1.10.22
sqlalchemy
psycopg2-binary
python-dotenv
